+++ crosswordPlayer/src/App.tsx
@@ -1,14 +1,8 @@
-import { Container, Box, Typography, Button, FormControl, InputLabel, Select, MenuItem, SelectChangeEvent } from '@mui/material';
+import { Container, Box, Typography, Button } from '@mui/material';
 import CrosswordBoard from './components/Crossword/CrosswordBoard';
 import CrosswordClues from './components/Crossword/CrosswordClues';
 import { useCrossword } from './hooks/useCrossword';
-import useLanguageGenerator from './hooks/useLanguageGenerator';
-import { useState, useEffect } from 'react';
-
-interface LanguageData {
-  languageCode: string;
-  projectName: string;
-}
+import LanguageSelector from './components/LanguageSelector';
 
 function App() {
   const { 
@@ -19,38 +13,6 @@ function App() {
     handleClick, 
     isActiveCell 
   } = useCrossword();
-
-  const { fetchLanguages, loading, error } = useLanguageGenerator(); 
-
-  const [menuLanguages, setMenuLanguages] = useState<LanguageData[]>([]);
-  const [selectedLanguage, setSelectedLanguage] = useState<string>("");
-
-  // Fetch languages on component mount
-  useEffect(() => {
-    const loadLanguages = async () => {
-      try {
-        const fetchedLanguages = await fetchLanguages();
-        console.log("Languages fetched:", fetchedLanguages);
-        if (fetchedLanguages) {
-          setMenuLanguages(fetchedLanguages);
-        }
-      } catch (err) {
-        console.error("Error fetching languages:", err);
-      }
-    };
-    
-    loadLanguages();
-  }, [fetchLanguages]);
-  
-  const handleLanguageChange = (event: SelectChangeEvent) => {
-    const value = event.target.value;
-    const selected = menuLanguages.find(lang => `${lang.projectName}-${lang.languageCode}` === value);
-    
-    if (selected) {
-      setSelectedLanguage(value);
-      console.log(`Selected project: ${selected.projectName}, language code: ${selected.languageCode}`);
-    }
-  };
 
   return (
     <Container 
@@ -66,39 +28,7 @@ function App() {
         pt: 4
       }}
     >
-      {/* Language selector */}
-      <Box sx={{ width: '100%', maxWidth: 600, mb: 3 }}>
-        <FormControl fullWidth size="small">
-          <InputLabel id="language-select-label">Select Language</InputLabel>
-          <Select
-            labelId="language-select-label"
-            id="language-select"
-            value={selectedLanguage}
-            label="Select Language"
-            onChange={handleLanguageChange}
-            disabled={loading}
-          >
-            {loading ? (
-              <MenuItem disabled>Loading languages...</MenuItem>
-            ) : error ? (
-              <MenuItem disabled>Error: {error}</MenuItem>
-            ) : menuLanguages.length === 0 ? (
-              <MenuItem disabled>No languages available</MenuItem>
-            ) : (
-              menuLanguages.map((language) => (
-                <MenuItem
-                  key={`${language.projectName}-${language.languageCode}`}
-                  value={`${language.projectName}-${language.languageCode}`}
-                >
-                  {language.projectName} ({language.languageCode})
-                </MenuItem>
-              ))
-            )}
-          </Select>
-        </FormControl>
-      </Box>
-
-      {/* Crossword puzzle */}
+      <LanguageSelector />
       <Box sx={{ 
         width: '100%',
         maxWidth: 600,
